MEGA refactor into separate modules
This is a huge architectural change that separates functionality into
two very defined layers:

 - client: handling the HTTP verbs, request / response parsing,
   cleaning, error handling, etc. Plus other GraphAPI low level features
   such as ``pagination`` and ``batch``.

 - api: the business layer that tries to simplify the access to Facebook
   GraphAPI.

NOTE: this commit potentially breaks **a lot** of backwards
compatibility. Plus, the documentation is NOT updated. This commit is
meant to be the starting point of the redesign. Once it is stable
enough, the documentation will be patched and updated.